requirements.txt
huggingface-hub==0.25.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
scikit-learn==1.3.2
